Add reusable logic for common artifact FS interactions
