+++ crates/resolute/src/error.rs
@@ -1,9 +1,10 @@
-use std::{fmt::Display, path::PathBuf};
-
 use reqwest::StatusCode;
 use semver::Version;
 
-use crate::mods::ResoluteMod;
+use crate::{
+	manager::artifacts::{ArtifactError, ArtifactErrorVec},
+	mods::ResoluteMod,
+};
 
 /// Error returned from a Downloader
 #[derive(thiserror::Error, Debug)]
@@ -46,10 +47,10 @@ pub enum Error {
 	ModNotInstalled(Box<ResoluteMod>),
 
 	#[error("artifact error: {0}")]
-	Artifact(ArtifactError),
+	Artifact(#[from] ArtifactError),
 
 	#[error("multiple artifact errors: {0}")]
-	Artifacts(ArtifactErrorVec),
+	Artifacts(#[from] ArtifactErrorVec),
 
 	#[error("no old artifact exists to delete")]
 	NoOldArtifact,
@@ -64,88 +65,6 @@ pub enum Error {
 	#[cfg(feature = "db")]
 	#[error("item not found in database: {0}")]
 	ItemNotFound(String),
-}
-
-/// An error performing an action on an artifact
-#[derive(Debug)]
-pub struct ArtifactError {
-	pub action: ArtifactAction,
-	pub path: Option<PathBuf>,
-	pub source: Box<Error>,
-}
-
-impl std::error::Error for ArtifactError {
-	fn source(&self) -> Option<&(dyn std::error::Error + 'static)> {
-		Some(&self.source)
-	}
-}
-
-impl Display for ArtifactError {
-	fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-		match &self.path {
-			Some(path) => write!(
-				f,
-				"artifact {} error for file ({}): {}",
-				self.action,
-				path.display(),
-				self.source
-			),
-			None => write!(f, "artifact error: {}", self.source),
-		}
-	}
-}
-
-/// An action being attempted on an artifact
-#[derive(Debug)]
-pub enum ArtifactAction {
-	Download,
-	Delete,
-	Rename,
-}
-
-impl Display for ArtifactAction {
-	fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-		let label = match self {
-			Self::Download => "download",
-			Self::Delete => "delete",
-			Self::Rename => "rename",
-		};
-
-		write!(f, "{}", label)
-	}
-}
-
-/// A Vec of artifact errors
-#[derive(Debug, Default)]
-pub struct ArtifactErrorVec(pub Vec<ArtifactError>);
-
-impl ArtifactErrorVec {
-	/// Creates a new empty error vec
-	pub fn new() -> Self {
-		Self(Vec::new())
-	}
-
-	#[inline]
-	pub fn push(&mut self, err: ArtifactError) {
-		self.0.push(err)
-	}
-
-	#[inline]
-	pub fn is_empty(&self) -> bool {
-		self.0.is_empty()
-	}
-
-	#[inline]
-	pub fn len(&self) -> usize {
-		self.0.len()
-	}
-}
-
-impl Display for ArtifactErrorVec {
-	fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-		let text = self.0.iter().map(|err| err.to_string()).collect::<Vec<_>>().join(", ");
-		write!(f, "[{}]", text)
-	}
 }
 
 /// Alias for a `Result` with the error type `resolute::Error`.
