+++ crates/resolute/src/manager/mod.rs
@@ -1,3 +1,4 @@
+pub(crate) mod artifacts;
 mod delete;
 mod download;
